Initial commit of the Program
This is the initial commit of the program. The goal of this project it to make
a 'skill' for the Amazon Echo that when called will return how many days until
a certain date or event. At this point only the 'calc_days.pl' script works
and events need to be manually added into the 'events.csv' file.