Have 'calc_days.pl' print out the difference in days
The only change in this one is to make the 'calc_days.pl' script now print out how many days are in between the asked days and the current day. This was not done before because I was not sure weather or not to make the script print the amount or return it.